blog: add code structure links to mlp post
This provides the foundation of the way in which we will be walking
through the basic data structures in the codebase. The idea is to
provide these first to anchor the implementations and make it clear what
state is needed and being managed in the implementation, which other
examples I had seen left for later, at the slight expense of overall
understanding in my experience.