app: embed CSP from file, bump to Go 1.16
This is intended to facilitate using the CSP in local development with
Hugo as well, having the same CSP both places. Go 1.16 introduces the
go:embed directive which will facilitate that.

The app/go.* file updates were the result of running `go mod tidy`.

While 1.16 is not the latest version of Go, it is the latest version
supported by Google App Engine standard runtimes [1] so that is what we
stick with for now. We can bump to Go 1.17 in the future.

[1] https://cloud.google.com/appengine/docs/standard/go/release-notes