utilize CSP in local development
This is intended to help eliminate issues where the CSP needs to be
updated for some website change, but the need for that change is missed
in local development until that change is deployed.

The CSP is being specified in a page meta tag [1] rather than a header
because I was not able to find an effective way to dynamically set a
value in Hugo at server run time. The environment variables [2] didn't
seem to work for this deeply nested map value, and the configDir
functionality [3] seemed overly complicated for this simple case, and
would have needed a non-standard configDir to be within app/ to place
nicely with Go embedding that did not seem warranted.

The self directive was added to connect-src for the local live reload
and should not cause any major concern for the deployed website.

[1] https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP
[2] https://gohugo.io/getting-started/configuration/#configure-with-environment-variables
[3] https://gohugo.io/getting-started/configuration/#configuration-directory