blog: update img shortcode for better positioning/sizing
This is aimed at better integrating the images with the post so have
things visually flow better. Also expands the shortcode so that the
figure element is used which can incorporate a caption, with a few
styles to match.

To help with these styles, we add some classes to the pages based on the
type and kind of the page from Hugo structure, replacing a hard-coded ID
which would also be incorrect in some cases.