img: switch homepage image styling to use more, taller version
This commit adds a new compressed version of the view-tall.jpg image
which was made with the "Export" function of apple preview, with a bit
of adjustment to bring the midpoint of the level adjustment downwards to
lighten up the darks a bit, putting a bit more focus on the text.

The CSS style updates are what cause the full image to be used, as well
as work better when at different webpage sizes and when re-sizing.