scripts: fix ordering bug in cleanup-appengine
This commit resolves a bug introduced into this script where the
ordering of the versions was using a non-existent variable name (though
it does appear in JSON output!) causing the newest branch builds to be
deleted rather than the oldest ones because the ordering was having no
effect, resulting in a situation where the thing that was just deployed
would then get deleted, but because of App Engine routing rules it would
just serve the default site version instead of erroring in some way,
resulting in confusion on my part as it seemed to be showing some
arbitrary old version.

The previous commits looking at CircleCI basically just proved that
CircleCI is working as expected, which led to looking at app engine and
finding this issue after observing that the version I thought I was
serving did not exist in the GCP console.

The App Engine behavior for falling back to the default version is
called "Soft Routing" and documented here:
https://cloud.google.com/appengine/docs/standard/go/how-requests-are-routed#soft_routing