blog: rename goml-mlp -> go-mlp
This slightly simpler name makes for a better URL I think, and now is
the time to change it before anything is "live".

Further, "goml" just happens to be the name of the repo, and the post is
about building an MLP in Go first and foremost, the fact that there is a
"goml" repo is secondary.