add robots.txt to disable crawling of branch builds
This makes a change so that builds to branches published to
app engine are not being indexed by crawlers in unwanted ways.

In order to facilitate this, I re-worked the build scripts
a bit to make the notion of an "environment" a first-class
thing with a separate script to define it based on the current
git branch.

References:
- Hugo robots page: https://gohugo.io/templates/robots/
- Hugo environments: https://gohugo.io/functions/hugo/
- Fun on comments in robots.txt: https://johnmu.com/robots-txt-comments/