remove reading time, word count from blog pages
Was pointed out in reviews that the reading time is not accurate for the
go-mlp post, and I think it is likely not accurate for any others as
well, particularly with technical material with diagrams. Word count
does not stand by itself really, so just removing this whole line for
now.