circleci: always deploy branch builds to App Engine
This simplifies the build logic and helps to handle cases where a branch
was pushed and then the PR opened later, where the CircleCI build will
not re-run in any capacity and thus leave the branch un-deployed. It is
simpler to have it just deploy all the time, and with the single-version
approach to naming the branches we are now using, it does not chew
through versions in the same way.