app: update zero-valued modification times on static files
This is a workaround for an issue seen with App Engine outlined here:
https://issuetracker.google.com/issues/168399701

When you upload files as part of the app to App Engine, it causes the
app to modify the timestamps of these files to match the value present
in the BUILD_DATE file, as though they had been generated at that time.

This is intended to work around issues observed at the current time
where the files are being cached for much longer than should be the case
due to these modification times in the headers set to values from 1980.

Also updates to the latest errors package.