app: embed site into binary, fixing modification timestamps
This commit resolves an issue where the Date associated with the files
on disk within App Engine is set to a non-zero but invalid time at the
beginning of 1980. By embedding the files, we drop the timestamps more
"correctly" acccording to how Go does things.

The hugo generated content directory had to be moved into the site/
folder for this since embedding cannot reach upwards for understandable
security reasons.

In future commits we may be able to bring back some of the previous
approach and inject the correct modification time into this static file
system by wrapping it.