index: move templating logic into scss file for jumbotron
While working on the search functionality I came across the "resource
from template" functionality that will be useful for generating the hugo
lunr index, which could apply to this hack done earlier to render and
compress the background image.

This change consolidates the CSS in the file that is now templated.