search: build index in JS, better JSON generation
This commit shifts the mechanism by which a JSON version of the site is
built by using the newer hugo resources capability for creating the
index data structure in-memory and then using the jsonify function to
render it out.

This switch in approach originally started out as a way to resolve an
issue with the quoting that changed behavior when the original script
was pulled into a js resource file. However, I actually like this
approach better.