search: switch to new preact JS lib for result rendering
This commit pulls in our first JS dependency for the site, used in a
targeted way to assist in the rendering of search results when a query
is dynamically executed rather than manually constructing the HTML text
from string concatenation.

Also pulls the "description" into the index format so that it is
available for rendering as well.