js: add jest test execution for search functionality
This commit brings in a number of changes to provide testing for the
search functionality in a very basic manner. The jest library is added
to provide the testing functionality, and we pull in lunr through node
modules now rather than a script so that it is available for use in the
tests.

To mirror the way that hugo does builds, we leverage the "esbuild"
integration with jest so that the same tooling is leveraged, with the
jest config file specifying how it is used.

One workaround needed which we can try to unwind in future commits is
that jQuery is not present so the functionality to listen for changes in
the search bar is not working.

Test execution is added to CircleCI as well.