search: add highlighting capabilities
This commit was inspired by the WIP post on the lunr index integration
into the site to add the ability to highlight matches in the text using
the basically undocumented functionality within Lunr to get this
information that was added in 2.0 but never fully worked into the docs.

Will update the post to incorporate this addition in follow up commits.