exp/playground: WASM-based playground up and running
This commit switches the strategy from one of gopherjs transpilation to
one where we are using the Go native WASM compilation target to expose a
function for template rendering.

This is utilized by some basic jQuery logic on the page to get the basic
functionality fully up and running!

I decided to switch away from gopherjs as it feels like it is not quite
needed for what we are doing here on the Go side, which is fairly basic.