circleci: switch Hugo install to binary release versions
This is intended to try to speed up the builds a bit, as the Go caching
did not seem to be working as expected consistently, e.g. [1] and it
makes it easier to be on a newer release version too.

The tools/ directory was removed accordingly since this was it's only
purpose, simplifying the repository structure. Locally I have been using
hugo from a package manager, so this was not seeing use there either.

[1] https://app.circleci.com/pipelines/github/kujenga/website/285/workflows/0a4bf4e6-92aa-491c-9ffa-edc50f3df95f/jobs/626