exp/go-templates: add support for Sprig functions library
This commit extends the playground to optionally (on by default) provide
the Sprig template functions library that is utilized by both Helm and
Hugo, extending the utility of this page.

The rather painful aspect of this is that it adds ~2MB to the size of
the WASM binary that is pulled in, due to the size of the Sprig library,
but I think it is worth it for the extra utility you get in being able
to work with additional scenarios where this extra functionality is an
integral part of working with templates.